# application/static/swagger.yaml
swagger: '2.0'
info:
  title: "Mechanic Shop API"
  description: "A comprehensive API for managing a mechanic shop including customers, mechanics, service tickets, and inventory"
  version: "1.0.0"
host: "127.0.0.1:5000"
schemes:
  - "http"
consumes:
  - "application/json"
produces:
  - "application/json"
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
    description: "JWT Authorization header using the Bearer scheme. Example: 'Bearer {token}'"
paths:
...
  /customers/login:
    post:
      tags:
        - Customers
      summary: "Customer login endpoint (generates token)"
      description: "Validates customer credentials, generates and returns a customer specific token."
      parameters:
        - in: "body"
          name: "body"
          description: "Login Credentials"
          required: true
          schema:
            $ref: "#/definitions/LoginCredentials"
      responses:
        200:
          description: "Login successful"
          schema:
            $ref: "#/definitions/LoginResponse"
          examples:
            application/json:
              message: "Login successful"
              token: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
              customer:
                id: 1
                first_name: "John"
                last_name: "Doe"
                email: "john.doe@email.com"
                phone: "555-1234"
                address: "123 Main St"
        400:
          description: "Invalid input"
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: "Invalid credentials"
          schema:
            $ref: "#/definitions/ErrorResponse"
  /customers:
    post:
      tags:
        - Customers
      summary: "Endpoint to create/register a new customer."
      description: "Endpoint to create a new customer in the Mechanic Shop system, emails must be unique."
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to create a new Customer."
          required: true
          schema:
            $ref: "#/definitions/CreateCustomerPayload"
      responses:
        201:
          description: "Successful Customer Creation"
          schema:
            $ref: "#/definitions/CustomerResponse"
          examples:
            application/json:
              id: 1
              first_name: "John"
              last_name: "Doe"
              email: "john.doe@email.com"
              phone: "555-1234"
              address: "123 Main St"
        400:
          description: "Invalid input"
          schema:
            $ref: "#/definitions/ErrorResponse"
        409:
          description: "Customer with this email already exists"
          schema:
            $ref: "#/definitions/ErrorResponse"

    get:
      tags:
        - Customers
      summary: "Returns all Customers"
      description: "Endpoint to retrieve a list of all customers."
      responses:
        200:
          description: "Retrieved Customers Successfully"
          schema:
            $ref: "#/definitions/AllCustomers"

    put:
      tags:
        - Customers
      summary: "Endpoint to update a customer."
      description: "Endpoint to update a customer, this is a token authenticated route, and the customer must be logged in to update their account."
      security:
        - bearerAuth: []
      parameters:
        - in: "body"
          name: "body"
          description: "Information required to update a Customer."
          required: true
          schema:
            $ref: "#/definitions/UpdateCustomerPayload"
      responses:
        200:
          description: "Successfully Updated customer"
          schema:
            $ref: "#/definitions/CustomerResponse"
          examples:
            application/json:
              id: 1
              first_name: "John"
              last_name: "Smith"
              email: "john.smith@email.com"
              phone: "555-5678"
              address: "456 Oak Ave"
        400:
          description: "Invalid input"
          schema:
            $ref: "#/definitions/ErrorResponse"
        401:
          description: "Unauthorized - Invalid or missing token"
          schema:
            $ref: "#/definitions/ErrorResponse"

    delete:
      tags:
        - Customers
      summary: "Endpoint to delete a customer"
      description: "Endpoint to delete a customer. Requires token authentication, customer must be logged in to delete their account"
      security:
        - bearerAuth: []
      responses:
        200:
          description: "Successfully Deleted customer"
          schema:
            $ref: "#/definitions/DeleteResponse"
          examples:
            application/json:
              message: "Customer successfully deleted"
        401:
          description: "Unauthorized - Invalid or missing token"
          schema:
            $ref: "#/definitions/ErrorResponse"
        404:
          description: "Customer not found"
          schema:
            $ref: "#/definitions/ErrorResponse"
  /customers/{customer_id}:
    get:
      tags:
        - Customers
      summary: "Get a specific customer by ID"
      description: "Retrieve detailed information about a specific customer"
      parameters:
        - in: "path"
          name: "customer_id"
          type: "integer"
          required: true
          description: "ID of the customer to retrieve"
      responses:
        200:
          description: "Customer retrieved successfully"
          schema:
            $ref: "#/definitions/CustomerResponse"
        404:
          description: "Customer not found"
          schema:
            $ref: "#/definitions/ErrorResponse"

  # Mechanic Endpoints
...
definitions:
...
  LoginCredentials:
    type: "object"
    properties:
      email:
        type: "string"
        format: "email"
        example: "john.doe@email.com"
      password:
        type: "string"
        format: "password"
        example: "SecurePassword123"
    required:
      - email
      - password
  LoginResponse:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Login successful"
      token:
        type: "string"
        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
      customer:
        type: "object"
        properties:
          id:
            type: "integer"
            example: 1
          first_name:
            type: "string"
            example: "John"
          last_name:
            type: "string"
            example: "Doe"
          email:
            type: "string"
            example: "john.doe@email.com"
          phone:
            type: "string"
            example: "555-1234"
          address:
            type: "string"
            example: "123 Main St"

  # Customer Definitions
  CreateCustomerPayload:
    type: "object"
    properties:
      first_name:
        type: "string"
        example: "John"
      last_name:
        type: "string"
        example: "Doe"
      email:
        type: "string"
        format: "email"
        example: "john.doe@email.com"
      phone:
        type: "string"
        example: "555-1234"
      address:
        type: "string"
        example: "123 Main St"
      password:
        type: "string"
        format: "password"
        example: "SecurePassword123"
    required:
      - first_name
      - last_name
      - email
      - password
  UpdateCustomerPayload:
    type: "object"
    properties:
      first_name:
        type: "string"
        example: "John"
      last_name:
        type: "string"
        example: "Smith"
      email:
        type: "string"
        format: "email"
        example: "john.smith@email.com"
      phone:
        type: "string"
        example: "555-5678"
      address:
        type: "string"
        example: "456 Oak Ave"
      password:
        type: "string"
        format: "password"
        example: "NewSecurePassword456"
    required:
      - first_name
      - last_name
      - email
  CustomerResponse:
    type: "object"
    properties:
      id:
        type: "integer"
        example: 1
      first_name:
        type: "string"
        example: "John"
      last_name:
        type: "string"
        example: "Doe"
      email:
        type: "string"
        example: "john.doe@email.com"
      phone:
        type: "string"
        example: "555-1234"
      address:
        type: "string"
        example: "123 Main St"
  AllCustomers:
    type: "array"
    items:
      type: "object"
      properties:
        id:
          type: "integer"
          example: 1
        first_name:
          type: "string"
          example: "John"
        last_name:
          type: "string"
          example: "Doe"
        email:
          type: "string"
          example: "john.doe@email.com"
        phone:
          type: "string"
          example: "555-1234"
        address:
          type: "string"
          example: "123 Main St"

  # Mechanic Definitions
...
  DeleteResponse:
    type: "object"
    properties:
      message:
        type: "string"
        example: "Successfully deleted"
  ErrorResponse:
    type: "object"
    properties:
      error:
        type: "string"
        example: "An error occurred"
      errors:
        type: "object"